refactor + add docs
